+++ src/Support/Utils.php
@@ -100,6 +100,24 @@ class Utils
     }
 
     /**
+     * Map a value or each value in an array.
+     *
+     * @param  mixed|array<mixed>  $valueOrValues
+     *
+     * @return mixed|array<mixed>
+     */
+    public static function mapEachRecursive(Closure $callback, $valueOrValues)
+    {
+        if (is_array($valueOrValues)) {
+            return array_map(function ($value) use ($callback) {
+                return static::mapEachRecursive($callback, $value);
+            }, $valueOrValues);
+        }
+
+        return $callback($valueOrValues);
+    }
+
+    /**
      * Apply a callback to a value or each value in an iterable.
      *
      * @param  mixed|iterable<mixed>  $valueOrValues