+++ tests/Unit/Schema/Types/LaravelEnumTypeTest.php
@@ -34,6 +34,19 @@ final class LaravelEnumTypeTest extends TestCase
         $this->assertSame($customName, $enumType->name);
     }
 
+    public function testEnumDescription(): void
+    {
+        $enumType = new LaravelEnumType(LocalizedUserType::class);
+
+        // TODO remove check when requiring bensampo/laravel-enum:6
+        // @phpstan-ignore-next-line depends on the required version
+        if (method_exists(LocalizedUserType::class, 'getClassDescription')) {
+            $this->assertSame('Localized user type', $enumType->config['description']);
+        } else {
+            $this->assertNull($enumType->config['description']);
+        }
+    }
+
     public function testCustomDescription(): void
     {
         $enumType = new LaravelEnumType(LocalizedUserType::class);
@@ -48,7 +61,11 @@ final class LaravelEnumTypeTest extends TestCase
     {
         $enumType = new LaravelEnumType(PartiallyDeprecated::class);
 
-        $this->assertSame(/** @lang GraphQL */ <<<GRAPHQL
+        // TODO remove check when requiring bensampo/laravel-enum:6
+        // @phpstan-ignore-next-line depends on the required version
+        if (method_exists(LocalizedUserType::class, 'getClassDescription')) {
+            $this->assertSame(/** @lang GraphQL */ <<<GRAPHQL
+"""Partially deprecated"""
 enum PartiallyDeprecated {
   "Not"
   NOT
@@ -60,9 +77,26 @@ enum PartiallyDeprecated {
   DEPRECATED_WITH_REASON @deprecated(reason: "some reason")
 }
 GRAPHQL
-            ,
-            SchemaPrinter::printType($enumType)
-        );
+                ,
+                SchemaPrinter::printType($enumType)
+            );
+        } else {
+            $this->assertSame(/** @lang GraphQL */ <<<GRAPHQL
+enum PartiallyDeprecated {
+  """Not"""
+  NOT
+
+  """Deprecated"""
+  DEPRECATED @deprecated
+
+  """Deprecated with reason"""
+  DEPRECATED_WITH_REASON @deprecated(reason: "some reason")
+}
+GRAPHQL
+                ,
+                SchemaPrinter::printType($enumType)
+            );
+        }
     }
 
     public function testReceivesEnumInstanceInternally(): void
@@ -103,5 +137,16 @@ GRAPHQL
         $this->expectExceptionObject(LaravelEnumType::classMustExtendBenSampoEnumEnum($notBenSampoEnumEnum));
         // @phpstan-ignore-next-line intentionally wrong
         new LaravelEnumType($notBenSampoEnumEnum);
+    }
+
+    public function testEnumMustHaveKey(): void
+    {
+        $enumType = new LaravelEnumType(AOrB::class);
+
+        $aWithoutKey = AOrB::A();
+        $aWithoutKey->key = null;
+
+        $this->expectExceptionObject(LaravelEnumType::enumMustHaveKey($aWithoutKey));
+        $enumType->serialize($aWithoutKey);
     }
 }